Handle errors in the main program
Errors are now handled in `main.rs` instead of `grm.rs`. It makes
testing easier and prepares the establishment of the lib. Grm will
have both a main program and a library.
Some required statements have been added to the CLI. Now help
is displayed when an argument is missing.